+++ common-functional/src/main/java/cn/featherfly/common/function/ToLocalDateTimeFunction.java
@@ -1,0 +1,15 @@
+package cn.featherfly.common.function;
+
+import java.time.LocalDateTime;
+import java.util.function.Function;
+
+/**
+ * The Interface ToLocalDateTimeFunction.
+ *
+ * @author zhongj
+ * @param <T> the type of the input to the function
+ */
+@FunctionalInterface
+public interface ToLocalDateTimeFunction<T> extends Function<T, LocalDateTime> {
+
+}