+++ common-functional/src/main/java/cn/featherfly/common/function/serializable/SerializableToDateFunction.java
@@ -2,14 +2,17 @@ package cn.featherfly.common.function.serializable;
 
 import java.util.Date;
 
+import cn.featherfly.common.function.ToDateFunction;
+
 /**
  * The Interface SerializableToDateFunction.
  *
  * @author zhongj
  * @param <T> the type of the input to the function
- * @param <R> the type of the result of the function
+ * @param <D> the type of the result of the function
  */
 @FunctionalInterface
-public interface SerializableToDateFunction<T, R extends Date> extends SerializableFunction<T, R> {
+public interface SerializableToDateFunction<T, D extends Date>
+        extends ToDateFunction<T, D>, SerializableFunction<T, D> {
 
 }