+++ common-functional/src/main/java/cn/featherfly/common/function/serializable/SerializableToEnumFunction.java
@@ -1,13 +1,16 @@
 package cn.featherfly.common.function.serializable;
 
+import cn.featherfly.common.function.ToEnumFunction;
+
 /**
  * The Interface ReturnEnumFunction.
  *
  * @author zhongj
  * @param <T> the type of the input to the function
- * @param <R> the type of the enum result of the function
+ * @param <E> the type of the enum result of the function
  */
 @FunctionalInterface
-public interface SerializableToEnumFunction<T, R extends Enum<R>> extends SerializableFunction<T, R> {
+public interface SerializableToEnumFunction<T, E extends Enum<E>>
+        extends ToEnumFunction<T, E>, SerializableFunction<T, E> {
 
 }