+++ common-functional/src/main/java/cn/featherfly/common/function/serializable/SerializableToLocalDateTimeFunction.java
@@ -2,6 +2,8 @@ package cn.featherfly.common.function.serializable;
 
 import java.time.LocalDateTime;
 
+import cn.featherfly.common.function.ToLocalDateTimeFunction;
+
 /**
  * The Interface SerializableToLocalDateTimeFunction.
  *
@@ -9,6 +11,7 @@ import java.time.LocalDateTime;
  * @param <T> the type of the input to the function
  */
 @FunctionalInterface
-public interface SerializableToLocalDateTimeFunction<T> extends SerializableFunction<T, LocalDateTime> {
+public interface SerializableToLocalDateTimeFunction<T>
+        extends ToLocalDateTimeFunction<T>, SerializableFunction<T, LocalDateTime> {
 
 }