+++ common-functional/src/main/java/cn/featherfly/common/function/serializable/SerializableToLocalTimeFunction.java
@@ -2,6 +2,8 @@ package cn.featherfly.common.function.serializable;
 
 import java.time.LocalTime;
 
+import cn.featherfly.common.function.ToLocalTimeFunction;
+
 /**
  * The Interface SerializableToLocalTimeFunction.
  *
@@ -9,6 +11,6 @@ import java.time.LocalTime;
  * @param <T> the type of the input to the function
  */
 @FunctionalInterface
-public interface SerializableToLocalTimeFunction<T> extends SerializableFunction<T, LocalTime> {
+public interface SerializableToLocalTimeFunction<T> extends ToLocalTimeFunction<T>, SerializableFunction<T, LocalTime> {
 
 }