+++ common-functional/src/main/java/cn/featherfly/common/function/serializable/SerializableToNumberFunction.java
@@ -1,13 +1,16 @@
 package cn.featherfly.common.function.serializable;
 
+import cn.featherfly.common.function.ToNumberFunction;
+
 /**
  * The Interface SerializableToNumberFunction.
  *
  * @author zhongj
  * @param <T> the type of the input to the function
- * @param <R> the type of the result of the function
+ * @param <N> the type of the result of the function
  */
 @FunctionalInterface
-public interface SerializableToNumberFunction<T, R extends Number> extends SerializableFunction<T, R> {
+public interface SerializableToNumberFunction<T, N extends Number>
+        extends ToNumberFunction<T, N>, SerializableFunction<T, N> {
 
 }