+++ common-functional/change-log.md
@@ -3,3 +3,4 @@
 2. 加入ArraySupplier<E>、ListSupplier<E>、SetSupplier<E>、QueueSupplier<E>、MapSupplier<K, V>等各种容器supplier
 3. 加入LocalXxxSupplier、StringSupplier等各种强类型supplier
 4. 加入各种Function，Consumer，Supplier对应的SerialzableXXX
+5. 加入ToLocalXxxFunction、ToStringFunction等各种强类型Function